app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy[asyncio]==1.4.46
asyncpg==0.27.0
alembic==1.11.1
pydantic==1.10.11
aiokafka==0.8.0
aioboto3==11.0.0
aioredis==2.0.2
motor==3.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.26.160
pydantic[email]==1.10.11
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.0
prometheus-client==0.16.0
gunicorn==21.2.0
python-dotenv==1.0.0
python-json-logger==2.0.4
